Added configurable regen, fixed patch compiler errors.
The largest part of this branch is the addition of the configurable regeneration framework, and all of its components.

I've also fixed a few errors that occurred due to issues in Harmony patches. In the ShouldHaveNeed patch, I changed it over to a void postfix, fixed the mis-named parameters, and accounted for pawns not having the IgnoreNeeds mod extension.

For the CostToMoveIntoCell patch, I fixed a few mis-named parameters and added the necessary annotations so that the proper overload is patched.

I should stress that, while these changes do fix the immediately obvious errors caused on load, I don't have any way to test if their intended functionality is still intact.