Fixed missing text in mod settings
Previously, the mod settings menu used field from the PatchSave class as the name and description for each patch toggle. However, these fields are never assigned to. Instead, I have refactored it to use the PatchSave.saveKey field to get the name and description when needed.

Edited the description for the caravan capacity patch to reflect its current function. Added some files to the gitignore. Re-included StatDefOf in the csproj, which was previously missing because of a fluke on my end.