Fixed side effects for missing parts
Fixed a bug in TryHealRandomPermanentWoundFor, which was caused by trying to add side effects to parts with Hediff_MissingPart hediffs. This was done by setting the missing part hediff's target part to null. An additional patch was needed to make this work without throwing an NRE.

I changed the save key for the missing parts patch, which highlighted a new issue: the mod settings still stores and displays save keys for patches that have been removed. I will look into this.