Exposed Output Workers, stability improvements
The base class for Output Workers is now public.  Added null check to PostProcessProduct's call, so that users can return no new products without throwing a NRE. Added null-or-empty checks to references to UseOutputWorkers.ActiveWorkers, for edge cases where the extension is used but no workers are provided. Added null-or-empty check to reference to UseOutputWorkers.outputWorkers for edge cases where the extension is used but no workers are provided. Updated manifest and assembly versions.